+++ src/pages/MyPage/components/SideNavigation/index.ts
@@ -1,0 +1,3 @@
+import SideNavigation from "./SideNavigation";
+
+export default SideNavigation;
